+++ gateway/src/apicast/configuration_loader/remote_v2.lua
@@ -18,6 +18,7 @@ local resty_env = require 'resty.env'
 local re = require 'ngx.re'
 local configuration = require 'apicast.configuration'
 local oidc_discovery = require('resty.oidc.discovery')
+local match = ngx.re.match
 
 local _M = {
   _VERSION = '0.1',
@@ -206,6 +207,15 @@ end
 function _M:call(environment)
   local load_just_for_host = load_just_the_services_needed()
 
+  local service_regexp_filter  = resty_env.value("APICAST_SERVICES_FILTER_BY_URL")
+  if service_regexp_filter then
+    local _, err = match("", service_regexp_filter, 'oj')
+    if err then
+      ngx.log(ngx.ERR, "APICAST_SERVICES_FILTER_BY_URL cannot compile, all services will be used: ", err)
+      service_regexp_filter = nil
+    end
+  end
+
   if self == _M  or not self then
     local host = environment
     local m = _M.new()
@@ -253,7 +263,7 @@ function _M:call(environment)
 
   local config
   for _, object in ipairs(res) do
-    config, err = self:config(object.service, env, 'latest')
+    config, err = self:config(object.service, env, 'latest', service_regexp_filter)
 
     if config then
       insert(configs, config)
@@ -338,7 +348,7 @@ function _M:oidc_issuer_configuration(service)
   return self.oidc:call(service.oidc.issuer_endpoint, self.ttl)
 end
 
-function _M:config(service, environment, version)
+function _M:config(service, environment, version, service_regexp_filter)
   local http_client = self.http_client
 
   if not http_client then return nil, 'not initialized' end
@@ -372,6 +382,10 @@ function _M:config(service, environment, version)
     local original_proxy_config = deepcopy(proxy_config)
 
     local config_service = configuration.parse_service(proxy_config.content)
+    if service_regexp_filter and not config_service:match_host(service_regexp_filter) then
+      return nil, "Service filtered out because APICAST_SERVICES_FILTER_BY_URL"
+    end
+
     original_proxy_config.oidc = self:oidc_issuer_configuration(config_service)
 
     return original_proxy_config